Add a getDepth() method to SubdivisionRepository.
Knowing how many levels of predefined subdivisions there are allows the form
builder to attach ajax callbacks only to the necessary subdivision fields.

This can't be determined simply by calling $subdivision->hasChildren() because
a single subdivision might not have children even though other subdivisions on
the same level do. Or, there might not be a subdivision selected yet (because the
field is optional).