Use ISO code as subdivision key.
In the interest of providing standardization to the data, this changes
the key in the subdivision files from Google's arbitrary strings to the
ISO ID for that subdivision, if found. Then, should there not be a
"name" value in the subdivision, if the Google "key" and ISO ID are not
the same it assigns the Google "key" as the name.